requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pandas==2.2.0
scikit-learn==1.4.0
websockets==12.0 